Simplify management of WAD file access
Do this by moving functions in w_file_stdc.c to the only place that uses
them, w_file.c.

Then recognize that the abstraction provided by wad_file_class_t has
only one existing implementation, the functions that we just moved to
w_file.c, and that any future new uses of this abstraction will better
go through a related similar we add to the `doomgeneric`` interface. So,
let's just delete all things related to wad_file_class_t and hardcode
the one implementation of it that exists.
