+++ doomgeneric/include/w_file.h
@@ -24,28 +24,7 @@
 
 typedef struct _wad_file_s wad_file_t;
 
-typedef struct {
-  // Open a file for reading.
-
-  wad_file_t *(*OpenFile)(char *path);
-
-  // Close the specified file.
-
-  void (*CloseFile)(wad_file_t *file);
-
-  // Read data from the specified position in the file into the
-  // provided buffer.  Returns the number of bytes read.
-
-  size_t (*Read)(wad_file_t *file, unsigned int offset, void *buffer,
-                 size_t buffer_len);
-
-} wad_file_class_t;
-
 struct _wad_file_s {
-  // Class of this file.
-
-  wad_file_class_t *file_class;
-
   // If this is NULL, the file cannot be mapped into memory.  If this
   // is non-NULL, it is a pointer to the mapped file.
