+++ doomgeneric/src/w_file.c
@@ -24,14 +24,69 @@
 #include "m_argv.h"
 
 #include "w_file.h"
+#include "m_misc.h"
+#include "z_zone.h"
 
-extern wad_file_class_t stdc_wad_file;
+typedef struct {
+  wad_file_t wad;
+  FILE *fstream;
+} stdc_wad_file_t;
 
-wad_file_t *W_OpenFile(char *path) { return stdc_wad_file.OpenFile(path); }
+static wad_file_t *W_StdC_OpenFile(char *path) {
+  stdc_wad_file_t *result;
+  FILE *fstream;
 
-void W_CloseFile(wad_file_t *wad) { wad->file_class->CloseFile(wad); }
+  fstream = fopen(path, "rb");
+
+  if (fstream == NULL) {
+    return NULL;
+  }
+
+  // Create a new stdc_wad_file_t to hold the file handle.
+
+  result = Z_Malloc(sizeof(stdc_wad_file_t), PU_STATIC, 0);
+  result->wad.mapped = NULL;
+  result->wad.length = M_FileLength(fstream);
+  result->fstream = fstream;
+
+  return &result->wad;
+}
+
+static void W_StdC_CloseFile(wad_file_t *wad) {
+  stdc_wad_file_t *stdc_wad;
+
+  stdc_wad = (stdc_wad_file_t *)wad;
+
+  fclose(stdc_wad->fstream);
+  Z_Free(stdc_wad);
+}
+
+// Read data from the specified position in the file into the
+// provided buffer.  Returns the number of bytes read.
+
+static size_t W_StdC_Read(wad_file_t *wad, unsigned int offset, void *buffer,
+                          size_t buffer_len) {
+  stdc_wad_file_t *stdc_wad;
+  size_t result;
+
+  stdc_wad = (stdc_wad_file_t *)wad;
+
+  // Jump to the specified position in the file.
+
+  fseek(stdc_wad->fstream, offset, SEEK_SET);
+
+  // Read into the buffer.
+
+  result = fread(buffer, 1, buffer_len, stdc_wad->fstream);
+
+  return result;
+}
+
+wad_file_t *W_OpenFile(char *path) { return W_StdC_OpenFile(path); }
+
+void W_CloseFile(wad_file_t *wad) { W_StdC_CloseFile(wad); }
 
 size_t W_Read(wad_file_t *wad, unsigned int offset, void *buffer,
               size_t buffer_len) {
-  return wad->file_class->Read(wad, offset, buffer, buffer_len);
+  return W_StdC_Read(wad, offset, buffer, buffer_len);
 }
